Refactor Beacon Inbox for performance and UI flow
- Extracted Beacon messaging logic to `BeaconPanel.jsx` to isolate state and prevent App-wide re-renders.
- Implemented `limit(50)` on message queries for faster initial load.
- Updated sorting logic to ensure newest threads are seen first in the Inbox.
- Implemented optimistic UI updates via Firestore metadata (`hasPendingWrites`) for instant "Sending" status.
- Added `slideInUp` animation and refined message bubble styling.
- Restored "Inbox lines" tabs functionality.
